Tried to fix things, but I'm not sure if th subject I have is the updated. I guess we have to wait for kick-off

+++ ft_strsub.c
@@ -23,7 +23,7 @@ char	*ft_strsub(char const *s, unsigned int start, size_t len)
 	str = ft_strnew(len + 1);
 	if (!str)
 		return (NULL);
-	while (++i < len + start)
+	while (++i < (int)len + (int)start)
 		str[i - start] = s[i];
 	return (str);
 }
